# backend/internal/server/openapi.yaml
openapi: 3.0.3
info:
  title: Gateway-LLM API
  description: |
    Unified LLM gateway that routes requests to OpenAI, Anthropic, and Gemini through a single
    OpenAI-compatible API. Provides API key management, team/org multi-tenancy, usage tracking,
    cost analytics, and role-based access control.

    ## Authentication

    All authenticated endpoints require a Bearer token in the `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    Tokens can be:
    - **Master key** — full admin access (set via `GATEWAY_LLM_MASTER_KEY`)
    - **API key** — created via `/v1/management/keys`, scoped to team/user/models
    - **Session token** — returned from `/v1/management/users/login`, expires in 24h

    ## Role Hierarchy

    `super_admin` > `org_admin` > `team_admin` > `member` > `viewer`
  version: 0.1.0
  contact:
    name: Gateway-LLM
    url: https://gateway-llm.com
servers:
  - url: https://api.gateway-llm.com
    description: Production
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Master key, API key, or session token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
...
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          description: Model alias configured in Gateway-LLM
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [system, user, assistant, function, tool]
              content:
                type: string
        temperature:
          type: number
        max_tokens:
          type: integer
        stream:
          type: boolean
        top_p:
          type: number
        frequency_penalty:
          type: number
        presence_penalty:
          type: number
        stop:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
              finish_reason:
                type: string
        usage:
          $ref: '#/components/schemas/Usage'
    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
    CompletionRequest:
      type: object
      required: [model, prompt]
      properties:
        model:
          type: string
        prompt:
          type: string
        max_tokens:
          type: integer
        temperature:
          type: number
        stream:
          type: boolean
    EmbeddingRequest:
      type: object
      required: [model, input]
      properties:
        model:
          type: string
        input:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        encoding_format:
          type: string
    ImageGenerationRequest:
      type: object
      required: [model, prompt]
      properties:
        model:
          type: string
        prompt:
          type: string
        n:
          type: integer
        size:
          type: string
    ModerationRequest:
      type: object
      required: [model, input]
      properties:
        model:
          type: string
        input:
          type: string
    ModelList:
      type: object
      properties:
        object:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              object:
                type: string
...
    ResponsesRequest:
      type: object
      required: [model, input]
      properties:
        model:
          type: string
        input:
          type: string
paths:
...
  /v1/chat/completions:
    post:
      summary: Create chat completion
      description: |
        OpenAI-compatible chat completion endpoint. Gateway-LLM routes the request to the
        appropriate provider based on the model alias. Supports streaming via `stream: true`.
      tags: [Chat]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
      responses:
        '200':
          description: Completion response
          headers:
            X-Gateway-LLM-Cost:
              schema:
                type: string
            X-Gateway-LLM-Tokens-Input:
              schema:
                type: string
            X-Gateway-LLM-Tokens-Output:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/completions:
    post:
      summary: Create text completion
      tags: [Completions]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompletionRequest'
      responses:
        '200':
          description: Completion response
        '401':
          description: Unauthorized
  /v1/embeddings:
    post:
      summary: Create embeddings
      tags: [Embeddings]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmbeddingRequest'
      responses:
        '200':
          description: Embedding response
        '401':
          description: Unauthorized
  /v1/images/generations:
    post:
      summary: Generate images
      tags: [Images]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageGenerationRequest'
      responses:
        '200':
          description: Image generation response
        '401':
          description: Unauthorized
  /v1/audio/speech:
    post:
      summary: Create speech
      description: Text-to-speech synthesis
      tags: [Audio]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, input, voice]
              properties:
                model:
                  type: string
                input:
                  type: string
                voice:
                  type: string
      responses:
        '200':
          description: Audio data
        '401':
          description: Unauthorized
  /v1/audio/transcriptions:
    post:
      summary: Transcribe audio
      description: Speech-to-text transcription
      tags: [Audio]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                model:
                  type: string
      responses:
        '200':
          description: Transcription result
        '401':
          description: Unauthorized
  /v1/moderations:
    post:
      summary: Create moderation
      tags: [Moderations]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerationRequest'
      responses:
        '200':
          description: Moderation result
        '401':
          description: Unauthorized
  /v1/responses:
    post:
      summary: Create response
      description: OpenAI Responses API endpoint
      tags: [Responses]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponsesRequest'
      responses:
        '200':
          description: Response object
        '401':
          description: Unauthorized
...
  /v1/models:
    get:
      summary: List models
      description: Returns all configured model aliases
      tags: [Models]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Model list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelList'
        '401':
          description: Unauthorized
...